backend/requirements.txt
absl-py
aiohttp
aiosignal
amqp==5.2.0
anyio
argon2-cffi
asgiref==3.7.2
asttokens
astunparse==1.6.3
async-lru
async-timeout
attrs
Babel
beautifulsoup4
billiard==4.2.0
bleach
blinker
Brotli
cachetools
celery==5.3.6
certifi
cffi
charset-normalizer
click
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama
comm
contourpy==1.2.0
cryptography
cycler==0.12.1
debugpy
decorator
defusedxml
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup
executing
fastjsonschema
flatbuffers
fonttools==4.47.2
frozenlist
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11==0.14.0
h5py
idna
imageio==2.33.1
imbalanced-learn==0.12.0
ipykernel
ipython
ipywidgets
jedi
Jinja2
joblib==1.3.2
json5
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keras
Keras-Preprocessing
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader
Markdown
MarkupSafe
matplotlib==3.8.2
matplotlib-inline
mistune
mkl-fft
mkl-random
mkl-service==2.4.0
multidict
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.2.1
notebook
notebook_shim
numpy
oauthlib
opencv-python==4.9.0.80
opt-einsum
outcome==1.3.0.post0
overrides
packaging
pandas==2.1.4
pandocfilters
parso
patsy==0.5.6
pillow==10.2.0
platformdirs
ply==3.11
prometheus-client
prompt-toolkit
protobuf==3.20.3
psutil
psycopg2==2.9.9
pure-eval
pyasn1
pyasn1-modules==0.2.8
pycparser
Pygments
PyJWT
pyOpenSSL
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip
PySocks
python-dateutil
python-json-logger
pytz
pywin32==305.1
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
requests-oauthlib==1.3.0
rfc3339-validator
rfc3986-validator
rpds-py
rsa
schedule==1.2.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.16.0
Send2Trash
sip
six
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve
sqlparse==0.4.4
stack-data
statsmodels==0.14.1
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow==2.10.0
tensorflow-estimator
termcolor
terminado
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2
tomli
tornado
traitlets
trio==0.24.0
trio-websocket==0.11.1
typing_extensions
tzdata==2023.4
urllib3
vine==5.1.0
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug
widgetsnbextension
win-inet-pton
wrapt
wsproto==1.2.0
yarl
